refactor(web): extract ChatMessageList and ChatComposer
Step 2 of 4 of the ChatForm refactor sequence. Pulls two pure-JSX
clusters out of ChatForm.tsx (453 → 253 LOC):

- ChatMessageList — chat log card + co-located NotionSaveRow.
  Props: messages, busy, retrying, notionReady, notionState, onNotionSave.
- ChatComposer — textarea + mic + send/cancel button cluster.
  Owns its own composingRef (IME guard) so ChatForm no longer juggles it.
  Props: input, setInput, busy, supportsMic, listening, onToggleMic,
  onSend, onCancel.

No behavior changes; all 85 tests still pass. ChatForm is now
orchestration-only (state plumbing + stream/send/cancel callbacks).
Step 3 will pull stream/send/cancel into a useChatStream hook.
